+++ src/main/java/org/modelingvalue/json/FromJsonGeneric.java
@@ -21,9 +21,7 @@ import java.util.Map;
 import java.util.Stack;
 import java.util.function.Consumer;
 
-import static org.modelingvalue.json.FromJsonGeneric.IdAcceptState.MAY_BE_ID;
-import static org.modelingvalue.json.FromJsonGeneric.IdAcceptState.MAY_BE_MORE;
-import static org.modelingvalue.json.FromJsonGeneric.IdAcceptState.MAY_NOT_BE_MORE;
+import static org.modelingvalue.json.FromJsonGeneric.IdAcceptState.*;
 
 public class FromJsonGeneric extends FromJsonBase<Object, Object> {
     public static <T> T fromJson(Type t, String s) {
@@ -54,23 +52,21 @@ public class FromJsonGeneric extends FromJsonBase<Object, Object> {
         pushType(t);
     }
 
-    private void pushType(Type subType) {
-        TypeInfo typeInfo = typeInfoMap.computeIfAbsent(subType, t -> TypeInfo.makeTypeInfo(t, config, topStackReplacer));
+    private void pushType(Type fieldType) {
+        TypeInfo typeInfo = typeInfoMap.computeIfAbsent(fieldType, t_ -> TypeInfo.makeTypeInfo(fieldType, config, topStackReplacer));
         typeInfoStack.push(typeInfo);
     }
 
     private Object makeObject() {
         Stack<Object> path = getPath();
         if (!path.isEmpty()) {
-            Type subType = typeInfoStack.peek().getSubType(path.peek());
-            pushType(subType);
+            pushType(typeInfoStack.peek().getFieldType(path.peek()));
         }
         return typeInfoStack.peek().getMaker().make();
     }
 
     private Object closeObject(Object m) {
-        typeInfoStack.pop();
-        return m;
+        return typeInfoStack.pop().convert(m);
     }
 
     ///////////////////////////////////////
@@ -87,30 +83,29 @@ public class FromJsonGeneric extends FromJsonBase<Object, Object> {
     protected Object makeMapEntry(Object m, Object key, Object value) {
         TypeInfo typeInfo = typeInfoStack.peek();
         switch (idAcceptState) {
-        case MAY_BE_ID:
-            idAcceptState = MAY_BE_MORE;
-            if (typeInfo.isIdField(key.toString())) {
-                final Object mm = m;
-                m = id2objectMap.computeIfAbsent(value, __ -> mm);
-                if (m != mm) {
-                    idAcceptState = MAY_NOT_BE_MORE;
+            case MAY_BE_ID:
+                idAcceptState = MAY_BE_MORE;
+                if (typeInfo.isIdField(key.toString())) {
+                    final Object mm = m;
+                    m = id2objectMap.computeIfAbsent(value, __ -> mm);
+                    if (m != mm) {
+                        idAcceptState = MAY_NOT_BE_MORE;
+                    }
                 }
-            }
-            break;
-        case MAY_BE_MORE:
-            if (typeInfo.isIdField(key.toString())) {
-                final Object mm = m;
-                m = id2objectMap.computeIfAbsent(value, __ -> mm);
-                if (m != mm) {
-                    throw error("id references must be the only field present when referencing a previous object: found " + key + ": " + value);
+                break;
+            case MAY_BE_MORE:
+                if (typeInfo.isIdField(key.toString())) {
+                    final Object mm = m;
+                    m = id2objectMap.computeIfAbsent(value, __ -> mm);
+                    if (m != mm) {
+                        throw error("id references must be the only field present when referencing a previous object: found " + key + ": " + value);
+                    }
                 }
-            }
-            break;
-        case MAY_NOT_BE_MORE:
-            throw error("id references must be the only field present when referencing a previous object: found " + key + ": " + value);
+                break;
+            case MAY_NOT_BE_MORE:
+                throw error("id references must be the only field present when referencing a previous object: found " + key + ": " + value);
         }
-        SubSetter subSetter = typeInfo.getSubSetter(key);
-        return subSetter.set(m, key, value);
+        return typeInfo.getFieldSetter(key).set(m, key, value);
     }
 
     @Override
@@ -127,8 +122,8 @@ public class FromJsonGeneric extends FromJsonBase<Object, Object> {
 
     @Override
     protected Object makeArrayEntry(Object a, int index, Object value) {
-        SubSetter subSetter = typeInfoStack.peek().getSubSetter(index);
-        return subSetter.set(a, index, value);
+        FieldSetter fieldSetter = typeInfoStack.peek().getFieldSetter(index);
+        return fieldSetter.set(a, index, value);
     }
 
     @Override
